# More Efficient `group_by(...) %>% sample_*(...)`